receiver/visualizer/requirements.txt
numpy==1.26.4
websockets
pyzmq